+++ core/server/api/v0.1/authentication.js
@@ -1,16 +1,16 @@
 const Promise = require('bluebird'),
     {extend, merge, omit, cloneDeep, assign} = require('lodash'),
     validator = require('validator'),
-    config = require('../config'),
-    common = require('../lib/common'),
-    security = require('../lib/security'),
-    constants = require('../lib/constants'),
-    pipeline = require('../lib/promise/pipeline'),
-    mail = require('../services/mail'),
-    urlService = require('../services/url'),
+    config = require('../../config'),
+    common = require('../../lib/common'),
+    security = require('../../lib/security'),
+    constants = require('../../lib/constants'),
+    pipeline = require('../../lib/promise/pipeline'),
+    mail = require('../../services/mail'),
+    urlService = require('../../services/url'),
     localUtils = require('./utils'),
-    models = require('../models'),
-    web = require('../web'),
+    models = require('../../models'),
+    web = require('../../web'),
     mailAPI = require('./mail'),
     settingsAPI = require('./settings'),
     tokenSecurity = {};
