+++ core/server/api/v0.1/invites.js
@@ -1,12 +1,12 @@
 const Promise = require('bluebird'),
     {omit, merge} = require('lodash'),
-    pipeline = require('../lib/promise/pipeline'),
-    mail = require('../services/mail'),
-    urlService = require('../services/url'),
+    pipeline = require('../../lib/promise/pipeline'),
+    mail = require('../../services/mail'),
+    urlService = require('../../services/url'),
     localUtils = require('./utils'),
-    models = require('../models'),
-    common = require('../lib/common'),
-    security = require('../lib/security'),
+    models = require('../../models'),
+    common = require('../../lib/common'),
+    security = require('../../lib/security'),
     mailAPI = require('./mail'),
     settingsAPI = require('./settings'),
     docName = 'invites',
