+++ core/server/api/v0.1/mail.js
@@ -2,11 +2,11 @@
 // API for sending Mail
 
 const Promise = require('bluebird'),
-    pipeline = require('../lib/promise/pipeline'),
+    pipeline = require('../../lib/promise/pipeline'),
     localUtils = require('./utils'),
-    models = require('../models'),
-    common = require('../lib/common'),
-    mail = require('../services/mail'),
+    models = require('../../models'),
+    common = require('../../lib/common'),
+    mail = require('../../services/mail'),
     notificationsAPI = require('./notifications'),
     docName = 'mail';
