+++ core/server/api/v0.1/slack.js
@@ -1,7 +1,7 @@
 // # Slack API
 // API for sending Test Notifications to Slack
 const Promise = require('bluebird'),
-    common = require('../lib/common');
+    common = require('../../lib/common');
 
 let slack;
