+++ core/server/api/v0.1/subscribers.js
@@ -3,11 +3,11 @@
 const Promise = require('bluebird'),
     _ = require('lodash'),
     fs = require('fs-extra'),
-    pipeline = require('../lib/promise/pipeline'),
-    fsLib = require('../lib/fs'),
+    pipeline = require('../../lib/promise/pipeline'),
+    fsLib = require('../../lib/fs'),
     localUtils = require('./utils'),
-    models = require('../models'),
-    common = require('../lib/common'),
+    models = require('../../models'),
+    common = require('../../lib/common'),
     docName = 'subscribers';
 
 let subscribers;
