+++ core/server/api/v0.1/tags.js
@@ -2,10 +2,10 @@
 // RESTful API for the Tag resource
 const Promise = require('bluebird'),
     _ = require('lodash'),
-    pipeline = require('../lib/promise/pipeline'),
+    pipeline = require('../../lib/promise/pipeline'),
     localUtils = require('./utils'),
-    models = require('../models'),
-    common = require('../lib/common'),
+    models = require('../../models'),
+    common = require('../../lib/common'),
     docName = 'tags',
     allowedIncludes = ['count.posts'];
