+++ core/server/api/v0.1/upload.js
@@ -1,6 +1,6 @@
 const Promise = require('bluebird'),
     fs = require('fs-extra'),
-    storage = require('../adapters/storage');
+    storage = require('../../adapters/storage');
 
 let upload;
