+++ core/test/functional/routes/dynamic_routing_spec.js
@@ -10,7 +10,7 @@ const should = require('should'),
     testUtils = require('../../utils'),
     cheerio = require('cheerio'),
     config = require('../../../../core/server/config'),
-    api = require('../../../../core/server/api'),
+    api = require('../../../server/api'),
     settingsCache = require('../../../server/services/settings/cache'),
     ghost = testUtils.startGhost,
     sandbox = sinon.sandbox.create();
