+++ core/test/integration/api/api_authentication_spec.js
@@ -3,8 +3,8 @@ var should = require('should'),
     testUtils = require('../../utils'),
     _ = require('lodash'),
     Promise = require('bluebird'),
-    AuthAPI = require('../../../server/api/authentication'),
-    mail = require('../../../server/api/mail'),
+    AuthAPI = require('../../../server/api/v0.1/authentication'),
+    mail = require('../../../server/api/v0.1/mail'),
     models = require('../../../server/models'),
     common = require('../../../server/lib/common'),
     security = require('../../../server/lib/security'),
