+++ core/test/integration/api/api_invites_spec.js
@@ -4,8 +4,8 @@ var should = require('should'),
     _ = require('lodash'),
     ObjectId = require('bson-objectid'),
     Promise = require('bluebird'),
-    InvitesAPI = require('../../../server/api/invites'),
-    mail = require('../../../server/api/mail'),
+    InvitesAPI = require('../../../server/api/v0.1/invites'),
+    mail = require('../../../server/api/v0.1/mail'),
     common = require('../../../server/lib/common'),
     context = testUtils.context,
