+++ core/test/integration/api/api_users_spec.js
@@ -5,7 +5,7 @@ var should = require('should'),
     _ = require('lodash'),
     models = require('../../../server/models'),
     common = require('../../../server/lib/common'),
-    UserAPI = require('../../../server/api/users'),
+    UserAPI = require('../../../server/api/v0.1/users'),
     db = require('../../../server/data/db'),
     context = testUtils.context,
     userIdFor = testUtils.users.ids,
